Visualizza mediacollection come utente normale

+++ private/mediacollection_p.h
@@ -22,7 +22,7 @@ public:
   Session *session;
   std::list<std::string> allowedPaths;
   Wt::WApplication *app;
-  long userId;
+  long long userId;
 };
 }
 #endif
